+++ test/fixtures/classname-summary/product/Name.php
@@ -1,0 +1,11 @@
+<?php
+namespace hoge\fuga\product;
+
+/**
+ * 製品名
+ * 
+ * 製品名のクラスです。
+ */
+class Name {
+    private string $name;
+}