+++ test/fixtures/classname-summary/product/Price.php
@@ -1,0 +1,11 @@
+<?php
+namespace hoge\fuga\product;
+
+/**
+ * 価格
+ * 
+ * 製品の価格です。
+ */
+class Price {
+    private int $price;
+}