+++ test/fixtures/classname-summary/product/Product.php
@@ -1,0 +1,21 @@
+<?php
+namespace hoge\fuga\product;
+
+use hoge\fuga\product\ {
+    Name,
+    Price,
+};
+
+/**
+ * 製品
+ * 
+ * 製品クラスです。
+ */
+class Product {
+    private Name $name;
+    private Price $price;
+
+    /** @return Product product */
+    public function method1(string $param1) {
+    }
+}